% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMCMR-defunct.R
\name{PMCMR-defunct}
\alias{PMCMR-defunct}
\alias{dunn.test.control}
\alias{jonckheere.test}
\alias{jonckheere.test.default}
\alias{posthoc.friedman.conover.test}
\alias{posthoc.friedman.conover.test.default}
\alias{posthoc.friedman.nemenyi.test}
\alias{posthoc.friedman.nemenyi.test.default}
\alias{posthoc.friedman.nemenyi.test.formula}
\alias{durbin.test}
\alias{durbin.test.default}
\alias{durbin.test.formula}
\alias{posthoc.kruskal.conover.test}
\alias{posthoc.kruskal.conover.test.default}
\alias{posthoc.kruskal.conover.test.formula}
\alias{posthoc.kruskal.dunn.test}
\alias{posthoc.kruskal.dunn.test.default}
\alias{posthoc.kruskal.dunn.test.formula}
\alias{posthoc.kruskal.nemenyi.test}
\alias{posthoc.kruskal.nemenyi.test.default}
\alias{posthoc.kruskal.nemenyi.test.formula}
\alias{posthoc.quade.test}
\alias{posthoc.quade.test.default}
\alias{posthoc.vanWaerden.test}
\alias{posthoc.vanWaerden.test.default}
\alias{posthoc.vanWaerden.test.formula}
\alias{vanWaerden.test}
\alias{vanWaerden.test.default}
\alias{vanWaerden.test.formula}
\title{PMCMR-defunct}
\usage{
dunn.test.control(x, g, p.adjust.method = p.adjust.methods, ...)

jonckheere.test(x, ...)

\method{jonckheere.test}{default}(
  x,
  g,
  alternative = c("monotonic", "increasing", "decreasing"),
  ...
)

posthoc.friedman.conover.test(y, ...)

\method{posthoc.friedman.conover.test}{default}(
  y,
  groups,
  blocks,
  p.adjust.method = p.adjust.methods,
  ...
)

posthoc.friedman.nemenyi.test(y, ...)

\method{posthoc.friedman.nemenyi.test}{default}(y, groups, blocks, ...)

\method{posthoc.friedman.nemenyi.test}{formula}(formula, data, subset, na.action, ...)

durbin.test(y, ...)

\method{durbin.test}{default}(y, groups, blocks, ...)

\method{durbin.test}{formula}(formula, data, subset, na.action, ...)

posthoc.kruskal.conover.test(x, ...)

\method{posthoc.kruskal.conover.test}{default}(x, g, p.adjust.method = p.adjust.methods, ...)

\method{posthoc.kruskal.conover.test}{formula}(
  formula,
  data,
  subset,
  na.action,
  p.adjust.method = p.adjust.methods,
  ...
)

posthoc.kruskal.dunn.test(x, ...)

\method{posthoc.kruskal.dunn.test}{default}(x, g, p.adjust.method = p.adjust.methods, ...)

\method{posthoc.kruskal.dunn.test}{formula}(
  formula,
  data,
  subset,
  na.action,
  p.adjust.method = p.adjust.methods,
  ...
)

posthoc.kruskal.nemenyi.test(x, ...)

\method{posthoc.kruskal.nemenyi.test}{default}(x, g, dist = c("Tukey", "Chisquare"), ...)

\method{posthoc.kruskal.nemenyi.test}{formula}(
  formula,
  data,
  subset,
  na.action,
  dist = c("Tukey", "Chisquare"),
  ...
)

posthoc.quade.test(y, ...)

\method{posthoc.quade.test}{default}(
  y,
  groups,
  blocks,
  dist = c("TDist", "Normal"),
  p.adjust.method = p.adjust.methods,
  ...
)

posthoc.vanWaerden.test(x, ...)

\method{posthoc.vanWaerden.test}{default}(x, g, p.adjust.method = p.adjust.methods, ...)

\method{posthoc.vanWaerden.test}{formula}(
  formula,
  data,
  subset,
  na.action,
  p.adjust.method = p.adjust.methods,
  ...
)

vanWaerden.test(x, ...)

\method{vanWaerden.test}{default}(x, g, ...)

\method{vanWaerden.test}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{a numeric vector of data values,
or a list of numeric data vectors.}

\item{g}{a vector or factor object giving
the group for the corresponding elements of \code{x}.
Ignored if \code{x} is a list.}

\item{p.adjust.method}{Method for adjusting
p values (see \code{\link[stats]{p.adjust}}).}

\item{\dots}{further arguments to be passed to or from methods.}

\item{alternative}{The alternative hypothesis.}

\item{y}{either a numeric vector of data values, or a data matrix.}

\item{groups}{a vector giving the group for
the corresponding elements of \code{y} if this is a vector;
ignored if \code{y} is a matrix. If not a factor object,
it is coerced to one.}

\item{blocks}{a vector giving the block for
the corresponding elements of \code{y} if this is a vector;
ignored if \code{y} is a matrix. If not a factor object,
it is coerced to one}

\item{formula}{a formula of the form \code{a ~ b | c},
where \code{a}, \code{b} and \code{c} give the data
values and corresponding groups and blocks, respectively.}

\item{data}{an optional matrix or data frame
(or similar: see \code{\link[stats]{model.frame}}) containing the
variables in the formula \code{formula}.
By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a
subset of observations to be used.}

\item{na.action}{a function which indicates what
should happen when the data contain \code{NA}s.
Defaults to \code{getOption("na.action")}.´}

\item{dist}{the test distribution}
}
\description{
The functions or methods listed here are no
longer part of \pkg{PMCMR}. You will find
functions and methods in the \pkg{PMCMRplus}
package \url{https://cran.r-project.org/package=PMCMRplus}.
}
\section{\code{\link{.Defunct}}}{
NA
}

