% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.durbin.test.R
\name{PMCMR-deprecated}
\alias{PMCMR-deprecated}
\alias{posthoc.durbin.test}
\alias{posthoc.durbin.test.default}
\title{Deprecated Functions in Package \pkg{PMCMR}}
\usage{
posthoc.durbin.test(y, ...)

\method{posthoc.durbin.test}{default}(y, groups, blocks, p.adjust.method = p.adjust.methods, ...)
}
\arguments{
\item{y}{either a numeric vector of data values,
or a data matrix.}

\item{\dots}{further arguments to be passed to or
from methods.}

\item{groups}{a vector giving the group for
the corresponding elements of \code{y} if this is a vector;
ignored if \code{y} is a matrix.
If not a factor object, it is coerced to one.}

\item{blocks}{a vector giving the block for the
corresponding elements of \code{y} if this is a vector;
ignored if \code{y} is a matrix. If not a factor object,
it is coerced to one.}

\item{p.adjust.method}{Method for adjusting p values
(see \code{\link[stats]{p.adjust}}).}
}
\value{
A list with class \code{"PMCMR"}
\itemize{
   \item{method }{The applied method.}
   \item{data.name}{The name of the data.}
   \item{p.value}{The two-sided p-value according to the student-t-distribution.}
   \item{statistic}{The estimated quantiles of the
       student-t-distribution.}
   \item{p.adjust.method}{The applied method for p-value adjustment.}
}
}
\description{
These functions are provided for reverse-dependencies
issues of other R-packages. They should no longer be used,
as actively maintained functions can be found in the package
\pkg{PMCMRplus}.
The functions may be defunct as soon as the next release.
}
\note{
The function does not test, whether it is a true BIBD.

This function does not test for ties.
}
\examples{

\dontrun{
## Example for an incomplete block design:
## Data from Conover (1999, p. 391).
y <- matrix(c(2, NA, NA, NA, 3, NA, 3, 3,
3, NA, NA, NA,  3, NA, NA,
1,  2, NA, NA, NA,  1,  1, NA,  1,  1,
NA, NA, NA, NA,  2, NA,  2,  1, NA, NA, NA, NA,
3, NA,  2,  1, NA, NA, NA, NA,  3, NA,  2,  2),
ncol=7, nrow=7, byrow=FALSE,
dimnames=list(1:7, LETTERS[1:7]))

posthoc.durbin.test(y)
}

}
\references{
W. J. Conover and R. L. Iman (1979), \emph{On multiple-comparisons
 procedures}, Tech. Rep. LA-7677-MS, Los Alamos Scientific Laboratory.

 W. J. Conover (1999), \emph{Practical nonparametric Statistics}, 3rd. Edition, Wiley.
}
