% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMCMR.R
\name{print.PMCMR}
\alias{print.PMCMR}
\title{Prints PMCMR objects}
\usage{
\method{print}{PMCMR}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"PMCMR"}, usually,
a result of a call to any of the posthoc-tests included
in the package PMCMR.}

\item{\ldots}{further arguments, currently ignored.}
}
\value{
The function \code{print.PMCMR} returns the lower
triangle of the (adjusted) p-values from any of
the posthoc tests included in the package PMCMR.
}
\description{
\code{print} method for class \code{"PMCMR"}.
}
