% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PMCMR.R
\name{summary.PMCMR}
\alias{summary.PMCMR}
\title{Summarizing PMCMR objects}
\usage{
\method{summary}{PMCMR}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"PMCMR"},
usually, a result of a call to any of the posthoc-tests
included in the package PMCMR.}

\item{\dots}{further arguments, currently ignored.}
}
\value{
The function \code{summary.PMCMR} computes and returns a list of the
pairwise comparisons including the H0, the corresponding statistic and
the (adjusted) p-value.
}
\description{
\code{summary} method for class \code{"PMCMR"}.
}
\keyword{methods}
\keyword{print}
